rasa_project/requirements.txt
asgiref==3.8.1
boto3==1.35.19
botocore==1.35.19
cffi==1.17.1
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dj-database-url==2.2.0
Django==5.0.7
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
fonttools==4.53.1
gunicorn==23.0.0
inflection==0.5.1
jmespath==1.0.1
kiwisolver==1.4.7
matplotlib==3.9.2
mysqlclient==2.2.4
numpy==2.1.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
PyAudio==0.2.14
pycparser==2.22
pydub==0.25.1
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
s3transfer==0.10.2
scipy==1.14.1
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.7.0
